requirements.txt
streamlit
google.generativeai
pypdf
opencv-python-headless
python-docx
pillow
python-dotenv